[Refactor] Create a UserController for the api endpoints in router.js
